fbscreencap:  Silence narrowing warnings

+++ graphics/fbscreencap/fbscreencap.cc
@@ -91,11 +91,11 @@ void save_desktop_screenshot( const char* path )
 	{
 		0,
 		0,
-		width,
-		height,
-		stride,
-		weight,
-		model,
+		(uint32_t) width,
+		(uint32_t) height,
+		(uint32_t) stride,
+		(uint8_t)  weight,
+		(uint8_t)  model,
 	};
 	
 	const bool rgba = is_rgba( var_info );
